Replace SingleModelCondition for DefaultSourcingCondition
Replace SingleModelCondition for DefaultSourcingCondition. Doing so
means we have an implementation that allows for several indices, as is
paramount of DCB. Furthermore, add a SourcingCondition#combine operation
 to merged conditions

#3010
